refactor(api/causes): modify causes create, update, list
