+++ src/main/java/com/linkedin/network/entity/Causes.java
@@ -1,8 +1,6 @@
-package com.linkedin.network.model;
+package com.linkedin.network.entity;
 
-import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
